+++ src/flockwave/channels/__init__.py
@@ -4,7 +4,13 @@ ReadableConnection_ and yields parsed message objects.
 
 from .encoder import EncoderChannel
 from .lossy import create_lossy_channel
-from .message import MessageChannel
+from .message import BroadcastMessageChannel, MessageChannel
 from .parser import ParserChannel
 
-__all__ = ("ParserChannel", "EncoderChannel", "MessageChannel", "create_lossy_channel")
+__all__ = (
+    "BroadcastMessageChannel",
+    "ParserChannel",
+    "EncoderChannel",
+    "MessageChannel",
+    "create_lossy_channel",
+)